requirements.txt
flask==2.3.2
firebase-admin==6.2.0
python-dotenv==1.0.0
requests==2.31.0
mailjet-rest==1.3.0
flask-uploads==0.2.1